Added prevention of inadvertent deletion of files that were not created by a file store's cache. See WWJ-469.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@1951 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWind/src/gov/nasa/worldwind/cache/AbstractFileStore.java
@@ -602,7 +602,8 @@ public abstract class AbstractFileStore extends WWObjectImpl implements FileStor
     }
 
     /**
-     * @param url the "file:" URL of the file to remove from the file store
+     * @param url the "file:" URL of the file to remove from the file store. Only files in the writable World Wind
+     *            disk cache or temp file directory are removed by this method.
      *
      * @throws IllegalArgumentException if <code>url</code> is null
      */
@@ -626,7 +627,15 @@ public abstract class AbstractFileStore extends WWObjectImpl implements FileStor
             synchronized (this.fileLock)
             {
                 if (file.exists())
+                {
+                    // Don't remove files outside the cache or temp directory.
+                    String parent = file.getParent();
+                    if (!(parent.startsWith(this.getWriteLocation().getPath())
+                        || parent.startsWith(Configuration.getSystemTempDirectory())))
+                        return;
+
                     file.delete();
+                }
             }
         }
         catch (java.net.URISyntaxException e)
